fix(helm): make upgrades more robust
The current way we were handling upgrades of Helm releases could
catch race conditions because we checked for status and then ran
the upgrade.  It's possible that both upgrades pass the status
check and then upgrade at the same time.

Instead, it's probably more wise to just do the upgrade and skip
failing if we get any error that the task is already happening
or that the release already exists.