fix(helm): skip deploying autoscaler unnecessarily
If there are any changes happening to node groups, we do not need
to deploy the Helm release for the auto scaler as there will not
be any changes to it.  This will reduce the possibility for any
race conditions happening with too many calls at once if there are
multiple node groups being created at once for example.