chore: add image builds inside ci
This commit will allow us to use "/build-new-image" in the body
of a PR and instead of downloading the latest shipped artifact or
image, it will build the image locally and test with it and then
publish it as a GitHub artifact.

Once we know it works, we can take that artifact and upload it
to the final object storage bucket for users to consume (or add
some automation which promotes the artifact after merging with
GitHub actions).