Add environment variable to set proxy port
The proxy selects a random port by default but this causes
difficulty in an environment where network traffic is controlled
by a firewall, either host iptables or a physical appliance.

This patch adds an environment variable, PROXY_PORT, which provides
a hint to select a preferred port, falling back to random selection
if that port is not available.