fix: stop unexpected rollouts (#324)
When making any node group changes, we were previously
making an entire change of all the cluster resources.  With
this change, we're instead simply making the small changes
in the node group which should avoid unexpected full rollouts
of node groups when changing min/max for autoscaling or
resizing a cluster.

Signed-off-by: Mohammed Naser <mnaser@vexxhost.com>