fix: force a cluster upgrade all the time
In the psat, if we get a cluster upgrade request when the kube_tag
did not change, we did not actually mutate the state of the cluster.

However, there is an expectation that an upgrade operation will
mutate the state of the cluster, so we will always apply the resources
when we see an upgrade come in.

Closes #381