Stop waiting for creation of node group
Since the update_nodegroup_status function is called for each
node group during updates, it will _not_ find the node group if
it hasn't been created yet, and it will keep the status as-is.

If it does find it, it will either find it in ScalingUp most likely, which
will then make the node group status get 'reset' to
CREATE_IN_PROGRESS and no harm done.

This eliminates the waiting that happens in the node group creation
part of the code.