Remove retries from delete node group
When deleting a node group, we can simply remove it and update the
status of it to DELETE_IN_PROGRESS, from there, the update node
group status code will check if it is progressing in the delete or not.

It will then clean up after it, so we can make that RPC call take less time.