Allow min_node_count to be zero
The current behavior sets min_node_count to 1 if the user specifies it
should be zero. Zero is supported by both Magnum and the Cluster
Autoscaler.

This sets it to the default value 1 (same as Magnum when min_node_count
is not specified) only when it was not specified.

Related:
* Magnum: https://review.opendev.org/c/openstack/magnum/+/737580
* Autoscaler: https://github.com/kubernetes/autoscaler/pull/3995