fix: use correct cloud_provider_tag (#532)
* reorder impls

* Initial resync to Helm chart

* add tests

* fix linter

* Write cloud controller manager config as 1001:1001

The Helm charts run the services on 1001:1001 and
therefore it can't read the file by default, instead of
switching it to run as root, instead we simply make
the file get written at 1001:1001 perms

Longer term we will move this into a secret instead.

* clean up resources.py

* Use secret for CCM

* Switch cluster resource to point to Secret

Since we're storing cloud-config inside, let's be clean
and store things in a secret.

* Switch cinder to use cloud-config

* Switch manila to use cloud-config secret

* Correct comment

* stop adding cloudcacert via clusterclass

* Drop CCM feature + switch to addon

* Use correct secret type